pyproject.toml
[project]
name = "xtend"
version = "0.1.0"
description = "Extend monitor via local web server"
authors = [
    {name = "Venkata Ramana",email = "pvrreddy155@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "aiofiles (==24.1.0)",
    "annotated-types (==0.7.0)",
    "anyio (==4.9.0)",
    "astroid (==3.3.9)",
    "certifi (==2025.1.31)",
    "cffi (==1.17.1)",
    "click (==8.1.8)",
    "dill (==0.3.9)",
    "fastapi (==0.115.12)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "iniconfig (==2.1.0)",
    "isort (==6.0.1)",
    "jinja2 (==3.1.6)",
    "loguru (==0.7.3)",
    "markupsafe (==3.0.2)",
    "mccabe (==0.7.0)",
    "mss (==10.0.0)",
    "numpy (==2.2.4)",
    "packaging (==24.2)",
    "platformdirs (==4.3.7)",
    "pluggy (==1.5.0)",
    "psutil (==7.0.0)",
    "pycparser (==2.22)",
    "pydantic (==2.10.6)",
    "pydantic-core (==2.27.2)",
    "pylint (==3.3.6)",
    "pytest (==8.3.5)",
    "python-multipart (==0.0.20)",
    "python-xlib (==0.33)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "starlette (==0.46.1)",
    "tomlkit (==0.13.2)",
    "typing-extensions (==4.12.2)",
    "uuid (==1.30)",
    "uvicorn (==0.34.0)",
    "websockets (==15.0.1)",
    "xcffib (==1.7.1)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pyside6 (>=6.9.0,<7.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]

[tool.poetry.scripts]
xtend = "xtend.cli:main"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
